refactor: drop unused plugins, edit file structure
