+++ src/components/Header.tsx
@@ -1,11 +1,10 @@
-import React, {useState, useRef} from 'react';
+import React from 'react';
 import IconButton from '@material-ui/core/IconButton';
 import {Logo} from './parts/Logo';
 import FullscreenIcon from '@material-ui/icons/Fullscreen';
 import VolumeOffIcon from '@material-ui/icons/VolumeOff';
 import VolumeUpIcon from '@material-ui/icons/VolumeUp';
-import MusicNoteIcon from '@material-ui/icons/MusicNote';
-import MusicOffIcon from '@material-ui/icons/MusicOff';
+import {BtnPlayMusic} from '../containers/BtnPlayMusic';
 
 interface IProps {
   onBtnFullScreenClick: () => void;
@@ -14,25 +13,6 @@ interface IProps {
 }
 
 export const Header: React.FC<IProps> = ({onBtnFullScreenClick, onBtnVolumeMuteClick, isMute}) => {
-  const [isMusicPlayed, setIsMusicPlayed] = useState<boolean>(true);
-  const refVideo = useRef<HTMLAudioElement>(null);
-
-  const handleMusicClick = () => {
-    setIsMusicPlayed(!isMusicPlayed);
-
-    if (!refVideo.current) {
-      return;
-    }
-
-    if (!isMusicPlayed) {
-      refVideo.current.defaultMuted = true;
-      refVideo.current.muted = true;
-    } else {
-      refVideo.current.defaultMuted = false;
-      refVideo.current.muted = false;
-    }
-  };
-
   return (
     <header>
       <Logo />
@@ -41,8 +21,7 @@ export const Header: React.FC<IProps> = ({onBtnFullScreenClick, onBtnVolumeMuteC
         <FullscreenIcon />
       </IconButton>
 
-      <IconButton onClick={handleMusicClick}>{!isMusicPlayed ? <MusicOffIcon /> : <MusicNoteIcon />}</IconButton>
-      {!isMusicPlayed && <audio autoPlay loop ref={refVideo} src="https://timra.ru/portfolio/audio/music.mp3" />}
+      <BtnPlayMusic />
 
       <IconButton onClick={onBtnVolumeMuteClick}>{!isMute ? <VolumeOffIcon /> : <VolumeUpIcon />}</IconButton>
     </header>
