--- src/redux/posts/index.ts
@@ -1,3 +1,0 @@
-export {default} from './reducers';
-export * from './actions';
-export * from './thunks';
